tests/requirements.txt
numpy
flask
flask-API
pandas
requests
Pillow
nanowire_flask==0.0.92
qrcode
pyzbar
